fix: typo + lint
