+++ src/app/view/editor-edit-tools-view-component/editor-edit-tools-view.component.ts
@@ -17,6 +17,7 @@ import {LabelGroupService} from "../../services/data/labelgroup.service";
 import {LabelGroup} from "../../models/labelGroup.model";
 import {VersionControlService} from "../../services/data/version-control.service";
 import {PositionTransformationService} from "../../services/util/position.transformation.service";
+import {AutoLayoutService} from "../../services/util/auto-layout.service";
 import {OrderingAlgorithm} from "../../data-structures/technical.data.structures";
 import {SbbRadioChange} from "@sbb-esta/angular/radio-button";
 
@@ -64,6 +65,7 @@ export class EditorEditToolsViewComponent implements OnDestroy {
     private uiInteractionService: UiInteractionService,
     private versionControlService: VersionControlService,
     private positionTransformationService: PositionTransformationService,
+    private autoLayoutService: AutoLayoutService,
   ) {
     this.nodeLabelGroups = this.labelGroupService.getLabelGroupsFromLabelRef(LabelRef.Node);
     this.trainrunLabelGroups = this.labelGroupService.getLabelGroupsFromLabelRef(LabelRef.Trainrun);
@@ -203,6 +205,14 @@ export class EditorEditToolsViewComponent implements OnDestroy {
 
   onAlignElementsBottom() {
     this.positionTransformationService.alignSelectedElementsToBottomBorder();
+  }
+
+  onInverseOptimizeLayout() {
+    this.autoLayoutService.optimizeLayout(true);
+  }
+
+  onOptimizeLayout() {
+    this.autoLayoutService.optimizeLayout(false);
   }
 
   private loadNetzgrafik(event: Event, callback: (dto: NetzgrafikDto) => void) {
