+++ src/assets/i18n/de.json
@@ -337,6 +337,11 @@
         "align-elements-right": "Rechtsbündig ausrichten",
         "align-elements-top": "Oben ausrichten",
         "align-elements-bottom": "Unten ausrichten",
+        "optimization-netzgrafik-title": "Netzgrafik-Layout - Optimierung",
+        "optimization-netzgrafik-tooltip": "Strecke die Netzgrafik, so dass keine Kante kürzer als die minimale Kantenlänge der Zugfahrten ist.",
+        "optimize-layout": "Strecke die Netzgrafik",
+        "inverse_optimization-netzgrafik-tooltip": "Komprimiere Netzgrafik soweit, dass keine Kante unnötigerweise länger als die minimale Kantelänge ist.",
+        "inverse_optimization-netzgrafik": "Komprimiere die Netzgrafik",
         "node-optimization-title": "Netzgrafik - Knotenoptimierung",
         "alphabeticalOrdering": "Alphabetisch",
         "alphabeticalOrderingTooltip": "Ports anhand der Zugkategorien alphabetisch anordnen.",
