+++ src/assets/i18n/en.json
@@ -337,6 +337,11 @@
         "align-elements-right": "Align right",
         "align-elements-top": "Align top",
         "align-elements-bottom": "Align bottom",
+        "optimization-netzgrafik-title": "Netzgrafik-Layout - Optimization",
+        "optimization-netzgrafik-tooltip": "Stretch the netzgrafik so that no edge is shorter than the minimum edge length of the trainruns.",
+        "optimize-layout": "Stretch the netzgrafik",
+        "inverse_optimization-netzgrafik-tooltip": "Compress the netzgrafik so that no edge is unnecessarily longer than the minimum edge length.",
+        "inverse_optimization-netzgrafik": "Compress the netzgrafik",
         "node-optimization-title": "Netzgrafik - Node optimization",
         "alphabeticalOrdering": "Alphabetical",
         "alphabeticalOrderingTooltip": "Order ports alphabetically by train categories.",
