+++ src/assets/i18n/fr.json
@@ -337,6 +337,11 @@
         "align-elements-right": "Aligner à droite",
         "align-elements-top": "Aligner le haut",
         "align-elements-bottom": "Aligner le bas",
+        "optimization-netzgrafik-title": "Netzgrafik-Layout - Optimisation",
+        "optimization-netzgrafik-tooltip": "Étirez le netzgrafik de sorte qu'aucun bord ne soit plus court que la longueur minimale des bords des trajets de train.",
+        "optimize-layout": "Étirez le netzgrafik",
+        "inverse_optimization-netzgrafik-tooltip": "Compressez le netzgrafik de sorte qu'aucun bord ne soit inutilement plus long que la longueur minimale des bords.",
+        "inverse_optimization-netzgrafik": "Compressez le netzgrafik",
         "node-optimization-title": "Réticulaire - Optimisation des noeuds",
         "alphabeticalOrdering": "Alphabétique",
         "alphabeticalOrderingTooltip": "Ordonne les ports par ordre alphabétique des catégories de train.",
